+++ src/app/(admin)/Sidebar.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/src/lib/utils";
+import { cn } from "@/lib/utils";
 import Link from "next/link";
 
 const links = [
