+++ src/app/(admin)/dashboard/page.tsx
