+++ src/app/(admin)/payments/DataTable.tsx
@@ -20,16 +20,16 @@ import {
   TableHead,
   TableHeader,
   TableRow,
-} from "@/src/components/ui/Table";
-import { Button } from "@/src/components/ui/Button";
-import { Input } from "@/src/components/ui/Input";
+} from "@/components/ui/Table";
+import { Button } from "@/components/ui/Button";
+import { Input } from "@/components/ui/Input";
 import {
   DropdownMenu,
   DropdownMenuCheckboxItem,
   DropdownMenuContent,
   DropdownMenuTrigger,
-} from "@/src/components/ui/DropdownMenu";
-import { DataTablePagination } from "@/src/components/ui/DataTable/Pagination";
+} from "@/components/ui/DropdownMenu";
+import { DataTablePagination } from "@/components/ui/DataTable/Pagination";
 interface DataTableProps<TData, TValue> {
   columns: ColumnDef<TData, TValue>[];
   data: TData[];
