+++ src/app/auth/page.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@/src/components/ui/Button";
+import { Button } from "@/components/ui/Button";
 import {
   Card,
   CardContent,
@@ -6,15 +6,10 @@ import {
   CardFooter,
   CardHeader,
   CardTitle,
-} from "@/src/components/ui/Card";
-import { Input } from "@/src/components/ui/Input";
-import { Label } from "@/src/components/ui/Label";
-import {
-  Tabs,
-  TabsContent,
-  TabsList,
-  TabsTrigger,
-} from "@/src/components/ui/Tabs";
+} from "@/components/ui/Card";
+import { Input } from "@/components/ui/Input";
+import { Label } from "@/components/ui/Label";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/Tabs";
 
 export default function Page() {
   return (
