+++ src/components/ui/DataTable/DataTableColumnHeader.tsx
@@ -6,15 +6,15 @@ import {
 } from "@radix-ui/react-icons";
 import { Column } from "@tanstack/react-table";
 
-import { cn } from "@/src/lib/utils";
-import { Button } from "@/src/components/ui/Button";
+import { cn } from "@/lib/utils";
+import { Button } from "@/components/ui/Button";
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuSeparator,
   DropdownMenuTrigger,
-} from "@/src/components/ui/DropdownMenu";
+} from "@/components/ui/DropdownMenu";
 
 interface DataTableColumnHeaderProps<TData, TValue>
   extends React.HTMLAttributes<HTMLDivElement> {
