+++ src/components/ui/DataTable/Pagination.tsx
@@ -6,14 +6,14 @@ import {
 } from "@radix-ui/react-icons";
 import { Table } from "@tanstack/react-table";
 
-import { Button } from "@/src/components/ui/Button";
+import { Button } from "@/components/ui/Button";
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from "@/src/components/ui/Select";
+} from "@/components/ui/Select";
 
 interface DataTablePaginationProps<TData> {
   table: Table<TData>;
